also cast to double inside denominator in squaring num_nodes so as not to overlfow implicit unsigned int which is only 32 bits in debug print for memusage

+++ src/digraph.c
@@ -1011,7 +1011,7 @@ digraph_t *load_digraph_from_arclist_file(FILE *pajek_file,
                         ((double)HASH_COUNT(g->mixTwoPathHashTab)*2*
                                  sizeof(twopath_record_t))/(1024*1024),
                         100*(double)HASH_COUNT(g->mixTwoPathHashTab) /
-                        (g->num_nodes*g->num_nodes),
+                        ((double)g->num_nodes*g->num_nodes),
                         ((double)sizeof(uint_t)*num_vertices*num_vertices) /
                         (1024*1024)));
   MEMUSAGE_DEBUG_PRINT(("MixTwoPath hash table overhead %f MB\n", (double)HASH_OVERHEAD(hh, g->mixTwoPathHashTab)/(1024*1024)));
@@ -1020,7 +1020,7 @@ digraph_t *load_digraph_from_arclist_file(FILE *pajek_file,
                         ((double)HASH_COUNT(g->inTwoPathHashTab)*
                                  (sizeof(twopath_record_t)))/(1024*1024),
                         100*(double)HASH_COUNT(g->inTwoPathHashTab) /
-                        (g->num_nodes*g->num_nodes),
+                        ((double)g->num_nodes*g->num_nodes),
                         ((double)sizeof(uint_t)*num_vertices*num_vertices) /
                         (1024*1024)));
   MEMUSAGE_DEBUG_PRINT(("InTwoPath hash table overhead %f MB\n", (double)HASH_OVERHEAD(hh, g->inTwoPathHashTab)/(1024*1024)));
@@ -1029,7 +1029,7 @@ digraph_t *load_digraph_from_arclist_file(FILE *pajek_file,
                         ((double)HASH_COUNT(g->outTwoPathHashTab)*2*
                                  sizeof(twopath_record_t))/(1024*1024),
                         100*(double)HASH_COUNT(g->outTwoPathHashTab) /
-                        (g->num_nodes*g->num_nodes),
+                        ((double)g->num_nodes*g->num_nodes),
                         ((double)sizeof(uint_t)*num_vertices*num_vertices) /
                         (1024*1024)));
   MEMUSAGE_DEBUG_PRINT(("OutTwoPath hash table overhead %f MB\n", (double)HASH_OVERHEAD(hh, g->outTwoPathHashTab)/(1024*1024)));
